Cosmic Dog
This game is called 'Cosmic Dog'. This little dog can be controlled by directional keys on keyboard.  You can get more score by striking the colorful ballon but, it should be careful if you hit the star because your body is going to be bigger. Because it could be harder to escape the sun. If you hit it, the game will be over.